style: 上传文件的一些微小的优化
